# src/swagger/openapi.yaml
openapi: 3.1.0
info:
  title: OnlyUsedTesla API
  version: 0.1.1
servers:
  - url: https://apiv2.onlyusedtesla.com
  - url: http://localhost:3003
security:
  - bearerAuth: []
...
paths:
...
  /sellers/{id}:
    get:
      tags: [Sellers]
      summary: Get seller by id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    patch:
      tags: [Sellers]
      summary: Update seller
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...